requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docutils==0.19
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
MarkupSafe==2.1.1
Pillow==9.3.0
pycodestyle==2.10.0
Pygments==2.13.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
